Add unseen handling
When starring is used to distinguish between interesting
and non-interesting articles, the assumption is that all headlines
are checked. This is not the case when there are a lot, and we just
mark them read without checking them. But then we should not consider
them for training. So add an "unseen" field to the database.
Articles before its introduction will have NULL in there. For new
articles it is "false" by default. In the case we mark unchecked
headlines, it will be "true". Add a button "Mark as unseen" that
marks all visible headlines as read and unseen.